Move PublishModal to its own directory
